+++ src/main/java/ClothesShop/Entity/MapperAdmin.java
@@ -9,7 +9,7 @@ public class MapperAdmin implements RowMapper<Admin> {
 	
 	public Admin mapRow(ResultSet rs, int rowNum)throws SQLException{
 		Admin admin = new Admin();
-		admin.setId_ad(rs.getLong("id_ad"));
+		admin.setId_ad(rs.getInt("id_ad"));
 		admin.setTaikhoan(rs.getString("taikhoan"));
 		admin.setPass(rs.getString("pass"));
 		admin.setHoten(rs.getString("hoten"));
